+++ src/shared/code-inspection/utils/codePatternHelper.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { DEFAULT_HOVER_KEY_PATTERNS } from '../constants';
 import { createCodePattern } from './codePatternExtractor';
-import { parseSheetNames } from './sheetNameParser';
+import { parseSheetNames } from '../../sheet-data/utils/sheetNameParser';
 
 export const getCodePattern = (): RegExp => {
 	const config = vscode.workspace.getConfiguration('languageHelper');