+++ src/shared/sheet-data/constants/index.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_TARGET_SHEET_NAMES = 'WD,ST,CD';