+++ src/shared/sheet-data/types/index.ts
@@ -1,7 +1,3 @@
-export type LanguageEntry = Record<string, string>;
-
-export type LanguageDictionary = Record<string, LanguageEntry>;
-
 export type GoogleSheetsApiResponse = {
 	values: string[][];
 	error?: {